// webapps/admin/generated/http-clients/backend/models/group-consent-status-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Shula API
 * API documentation for the Shula application
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface GroupConsentStatusResponseDto
 */
export interface GroupConsentStatusResponseDto {
    /**
     * Group ID
     * @type {string}
     * @memberof GroupConsentStatusResponseDto
     */
    'groupId': string;
    /**
     * Whether the group has enough consents to be considered approved
     * @type {object}
     * @memberof GroupConsentStatusResponseDto
     */
    'isApproved': object | null;
    /**
     * Group consent status as a string
     * @type {string}
     * @memberof GroupConsentStatusResponseDto
     */
    'status': GroupConsentStatusResponseDtoStatusEnum;
}

export const GroupConsentStatusResponseDtoStatusEnum = {
    Approved: 'approved',
    Rejected: 'rejected',
    Pending: 'pending'
} as const;

export type GroupConsentStatusResponseDtoStatusEnum = typeof GroupConsentStatusResponseDtoStatusEnum[keyof typeof GroupConsentStatusResponseDtoStatusEnum];