// webapps/admin/generated/http-clients/backend/models/participant-consent-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Shula API
 * API documentation for the Shula application
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ParticipantConsentResponseDto
 */
export interface ParticipantConsentResponseDto {
    /**
     * Consent record ID
     * @type {string}
     * @memberof ParticipantConsentResponseDto
     */
    'id': string;
    /**
     * Group ID
     * @type {string}
     * @memberof ParticipantConsentResponseDto
     */
    'groupId': string;
    /**
     * Participant ID
     * @type {string}
     * @memberof ParticipantConsentResponseDto
     */
    'participantId': string;
    /**
     * Consent status
     * @type {string}
     * @memberof ParticipantConsentResponseDto
     */
    'consentStatus': ParticipantConsentResponseDtoConsentStatusEnum;
    /**
     * Date when the participant responded
     * @type {string}
     * @memberof ParticipantConsentResponseDto
     */
    'respondedAt'?: string | null;
    /**
     * Participant name
     * @type {string}
     * @memberof ParticipantConsentResponseDto
     */
    'participantName'?: string;
    /**
     * Participant phone number
     * @type {string}
     * @memberof ParticipantConsentResponseDto
     */
    'participantPhoneNumber'?: string;
}

export const ParticipantConsentResponseDtoConsentStatusEnum = {
    Pending: 'pending',
    Accepted: 'accepted',
    Rejected: 'rejected'
} as const;

export type ParticipantConsentResponseDtoConsentStatusEnum = typeof ParticipantConsentResponseDtoConsentStatusEnum[keyof typeof ParticipantConsentResponseDtoConsentStatusEnum];