// webapps/admin/generated/http-clients/backend/models/update-participant-consent-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Shula API
 * API documentation for the Shula application
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateParticipantConsentDto
 */
export interface UpdateParticipantConsentDto {
    /**
     * Consent status
     * @type {string}
     * @memberof UpdateParticipantConsentDto
     */
    'consentStatus': UpdateParticipantConsentDtoConsentStatusEnum;
}

export const UpdateParticipantConsentDtoConsentStatusEnum = {
    Pending: 'pending',
    Accepted: 'accepted',
    Rejected: 'rejected'
} as const;

export type UpdateParticipantConsentDtoConsentStatusEnum = typeof UpdateParticipantConsentDtoConsentStatusEnum[keyof typeof UpdateParticipantConsentDtoConsentStatusEnum];